Add Mouse as Touch and Mouse Gesture support. The behavior is off by default and is controlled via TouchPanel.EnableMouseTouchPoint and TouchPanel.EnableMouseGestures It is only supported on Windows 8 with this check in.

+++ MonoGame.Framework/Input/Touch/TouchPanel.cs
@@ -96,8 +96,9 @@ namespace Microsoft.Xna.Framework.Input.Touch
 
         /// <summary>
         /// The next touch location identifier.
+        /// The value 1 is reserved for the mouse touch point.
         /// </summary>
-        private static int _nextTouchId = 1;
+        private static int _nextTouchId = 2;
 
         /// <summary>
         /// The mapping between platform specific touch ids
@@ -202,6 +203,11 @@ namespace Microsoft.Xna.Framework.Input.Touch
 
         internal static void AddEvent(int id, TouchLocationState state, Vector2 position)
         {
+            AddEvent(id, state, position, true);
+        }
+
+        internal static void AddEvent(int id, TouchLocationState state, Vector2 position, bool isTouch)
+        {
             // Different platforms return different touch identifiers
             // based on the specifics of their implementation and the
             // system drivers.
@@ -215,7 +221,17 @@ namespace Microsoft.Xna.Framework.Input.Touch
             // and release events.
             // 
             if (state == TouchLocationState.Pressed)
-                _touchIds[id] = _nextTouchId++;
+            {
+                if (isTouch)
+                {
+                    _touchIds[id] = _nextTouchId++;
+                }
+                else
+                {
+                    // Mouse pointing devices always use touchId of 1
+                    _touchIds[id] = 1;
+                }
+            }
 
             // Try to find the touch id.
             int touchId;
@@ -227,20 +243,27 @@ namespace Microsoft.Xna.Framework.Input.Touch
                 return;
             }
 
-            // Add the new touch event keeping the list from getting
-            // too large if no one happens to be requesting the state.
-            var evt = new TouchLocation(touchId, state, position*_touchScale);
-            _touchEvents.Add(evt);
-            if (_touchEvents.Count > MaxEvents)
-                _touchEvents.RemoveRange(0, _touchEvents.Count - MaxEvents);
-
-            // If we have gestures enabled then start to collect 
-            // events for those too.
-            if (EnabledGestures != GestureType.None)
+            if (isTouch || EnableMouseTouchPoint || EnableMouseGestures)
             {
-                _gestureEvents.Add(evt);
-                if (_gestureEvents.Count > MaxEvents)
-                    _gestureEvents.RemoveRange(0, _gestureEvents.Count - MaxEvents);
+                // Add the new touch event keeping the list from getting
+                // too large if no one happens to be requesting the state.
+                var evt = new TouchLocation(touchId, state, position * _touchScale);
+
+                if (isTouch || EnableMouseTouchPoint)
+                {
+                    _touchEvents.Add(evt);
+                    if (_touchEvents.Count > MaxEvents)
+                        _touchEvents.RemoveRange(0, _touchEvents.Count - MaxEvents);
+                }
+
+                // If we have gestures enabled then start to collect 
+                // events for those too.
+                if (EnabledGestures != GestureType.None && (isTouch || EnableMouseGestures))
+                {
+                    _gestureEvents.Add(evt);
+                    if (_gestureEvents.Count > MaxEvents)
+                        _gestureEvents.RemoveRange(0, _gestureEvents.Count - MaxEvents);
+                }
             }
 
             // If this is a release unmap the hardware id.
@@ -338,6 +361,10 @@ namespace Microsoft.Xna.Framework.Input.Touch
         }
 		
         public static GestureType EnabledGestures { get; set; }
+
+        public static bool EnableMouseTouchPoint { get; set; }
+
+        public static bool EnableMouseGestures { get; set; }
 
         public static bool IsGestureAvailable
         {
